[hotfix] Move Bucketer interface to file sink base package
This brings it in line with RollingPolicy.
